requirements.txt
flask
flask-cors
paramiko
python-dotenv
cryptography
Flask-Login  
werkzeug  